refactor: extract game types to src/types/GameTypes.ts (step 1 of 6)
Move all exported interfaces and the GameSource type out of Game.ts into
a dedicated GameTypes.ts so extracted modules no longer import types
from the class file. Update all 14 external call sites and the 4 modules
from PR #910.
