+++ src/commands/admin/gotm-audit-handlers.ts
@@ -16,7 +16,7 @@ import {
   replyIfNotOwner,
 } from "../../functions/InteractionUtils.js";
 import { buildComponentsV2Flags, buildTextReply } from "../../functions/ComponentsV2Utils.js";
-import type { IGameWithPlatforms } from "../../classes/Game.js";
+import type { IGameWithPlatforms } from "../../types/GameTypes.js";
 import Game from "../../classes/Game.js";
 import {
   getGotmAuditImportById,
