+++ src/commands/gamedb/gamedb-profile.service.ts
@@ -24,7 +24,8 @@ import { formatPlatformDisplayName } from "../../functions/PlatformDisplay.js";
 import { formatTableDate } from "../../functions/DateFormatUtils.js";
 import { renderUsernameWithEmoji } from "../../services/UserEmojiService.js";
 import { padCommandName } from "../help.command.js";
-import Game, { type IGame, type IRelease } from "../../classes/Game.js";
+import type { IGame, IRelease } from "../../types/GameTypes.js";
+import Game from "../../classes/Game.js";
 import {
   buildComponentsV2Flags,
   getSearchRowsFromComponents,
