+++ src/commands/now-playing.command.ts
@@ -33,7 +33,8 @@ import {
   type AnyRepliable,
   replyIfNotOwner,
 } from "../functions/InteractionUtils.js";
-import Game, { type IGame } from "../classes/Game.js";
+import type { IGame } from "../types/GameTypes.js";
+import Game from "../classes/Game.js";
 import {
   buildActionButton,
   buildButtonRow,
